requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
textblob>=0.17.0
python-dotenv>=1.0.0
